Moved prep to Ecf4Helper from ServiceHelpers
As it specifically handles ECF4 QueryMessages.
